collecte/requirements.txt
kafka-python
pandas
faker
cassandra-driver
python-dotenv
requests